feat: remove parser
