--- src/app/app.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-
-@Module({
-  controllers: [AppController],
-  providers: [AppService]
-})
-export class AppModule {}
