--- src/app/app.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AppService } from './app.service';
-
-describe('AppService', () => {
-  let service: AppService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [AppService],
-    }).compile();
-
-    service = module.get<AppService>(AppService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
